Microservices/k8Azure/Payment/requirements.txt
flask_sqlalchemy
flask_cors
pika
pyodbc
stripe
